+++ Source/CDR.DataHolder.API.Logger/IRequestResponseLogger.cs
@@ -1,0 +1,9 @@
+﻿namespace CDR.DataHolder.API.Logger
+{
+    using Serilog;
+
+    public interface IRequestResponseLogger
+    {
+        ILogger Log { get; }
+    }
+}