+++ Source/CDR.DataHolder.Resource.API/Startup.cs
@@ -4,6 +4,7 @@ using CDR.DataHolder.API.Infrastructure.Filters;
 using CDR.DataHolder.API.Infrastructure.IdPermanence;
 using CDR.DataHolder.API.Infrastructure.Middleware;
 using CDR.DataHolder.API.Infrastructure.Models;
+using CDR.DataHolder.API.Logger;
 using CDR.DataHolder.Domain.Repositories;
 using CDR.DataHolder.Repository;
 using CDR.DataHolder.Repository.Infrastructure;
@@ -87,6 +88,12 @@ namespace CDR.DataHolder.Resource.API
             services.AddAutoMapper(typeof(Startup), typeof(DataHolderDatabaseContext));
 
             services.AddScoped<LogActionEntryAttribute>();
+
+            if (Configuration.GetSection("SerilogRequestResponseLogger") != null)
+            {
+                Log.Logger.Information("Adding request response logging middleware");
+                services.AddRequestResponseLogging();
+            }
         }
 
         private static void AddAuthenticationAuthorization(IServiceCollection services, IConfiguration configuration)
@@ -174,6 +181,7 @@ namespace CDR.DataHolder.Resource.API
                     }
                 });
             });
+
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -186,8 +194,11 @@ namespace CDR.DataHolder.Resource.API
 
             app.UseSerilogRequestLogging();
 
+            app.UseMiddleware<RequestResponseLoggingMiddleware>();
             // ExceptionHandlingMiddleware must be first in the line, so it will catch all unhandled exceptions.
             app.UseMiddleware<ResourceAuthoriseErrorHandlingMiddleware>();
+            
+
 
             app.UseSwagger();
             app.UseSwaggerUI(c => c.SwaggerEndpoint("/swagger/v1/swagger.json", "Mock Data Holder Discovery API v1"));
@@ -198,6 +209,8 @@ namespace CDR.DataHolder.Resource.API
 
             app.UseAuthentication();
             app.UseAuthorization();
+
+            
 
             // Add custom middleware
             app.UseInteractionId();